ai-service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
insightface
onnxruntime
Pillow
prometheus-fastapi-instrumentator
